Add assembler implementation using PEXT
The integration with the benchmark/test code is a horrible hack, I know.

This fails one test, which is asserting that, in the case of bogus
padding in a multi-byte sequence, we skip the entire sequence. However,
this is a poorly defined case - the tests want us to skip the first byte
(only) if the first byte has bad padding, but the entire sequence if the
first byte looks okay. This is a bit awkward to express; I chose to
simply skip one byte forward in the case of bad UTF8 padding regardless
of which byte had the bad padding.

Note that this is not truly 'branchless', as it has branches that are
taken for invalid UTF-8 sequences. However, since these branches should
be trivially predictable, I would expect the performance to be similar
or better than a true branchless implementation.

Benchmark results on an EC2 c4.xlarge:

asm: 577.333402 MB/s, 0 errors
branchless: 410.666716 MB/s, 0 errors
Hoehrmann:  344.000041 MB/s, 0 errors